Handle multiple manifests in HLS LLStorage
This will allow for storing and serving multi-bitrate adaptive HLS

Manifest names are used as the namespace for tracking sequence numbers, etc, and
are being derived from (partial-)segment names. This shouldn't be a
problem as the segment names seem to be standarized, but there might be
a better way to map Segments to Manifests.